Added Documentation.md (See Description)
hi, all! this is a rough draft of the documentation for bender with sections to understand the current capability, and flaws. this is only a rough draft, and took a very long time to create, so some aspects may not be fully understandable. however, it will be updated as frequent as possible.

thanks! 💝