Added changes to reflect the new Documentation File
hi! i added the following changes to bender:

* two new customizable variables, called "ChatlogsSpamText", and "ChatSpamText". 

"ChatlogsSpamText" allows you to modify the command that is used to spam the in-game logs. make sure that it's a command though or it will not work
"ChatSpamText" allows you to change the text that is spammed in the in-game textbox. this, however, doesnt have to be a command though

good night guys im so tired