Added a new playerlist to reflect upcoming updates (SEE DESCRIPTION)
hi, everyone! this update includes a new playerlist that informs you if someone has perm, or person299.

the playerlist displays the user's full username, and will occasionally refresh itself to reflect any changes to the user count. 

therefore, it may take a few seconds to request the roblox api for this information, to fully refresh the playerlist.